+++ milvus/client/asynch.py
@@ -43,11 +43,13 @@ class Future(AbstractFuture):
         self._canceled = False
         self._done = False
         self._response = None
+        self._exception = None
 
         self.__init()
 
     def __del__(self):
-        pass
+        self._future.__del__()
+        self._future = None
         # self._future.__del__()
 
     @abc.abstractmethod
@@ -64,21 +66,21 @@ class Future(AbstractFuture):
 
         def async_done_callback(future):
             with self._condition:
-                self._response = future.result()
                 # delete gRCP future manually
-                self._future.__del__()
-                self._future = None
+                # self._future.__del__()
+                # self._future = None
 
                 # If user specify done callback function, execute it.
                 try:
+                    self._response = future.result()
                     if self._done_cb:
                         results = self.on_response(self._response)
                         if isinstance(results, tuple):
                             self._done_cb(*self.on_response(self._response))
                         else:
                             self._done_cb(self.on_response(self._response))
-                except:
-                    raise
+                except Exception as e:
+                    self._exception = e
                 finally:
                     self._done = True
                     self._condition.notify_all()
@@ -86,16 +88,18 @@ class Future(AbstractFuture):
         self._future.add_done_callback(async_done_callback)
 
     def result(self, **kwargs):
-
+        self.exception()
         with self._condition:
             # future not finished. wait callback being called.
-            if not self._done and not self._canceled:
-                to = kwargs.get("timeout", None)
-                self._condition.wait(to)
-
-                if not self._done and not self._canceled:
-                    # self._condition.notify_all()
-                    raise FutureTimeoutError("Wait timeout")
+            to = kwargs.get("timeout", None)
+            self._response = self._future.result(timeout=to)
+            # if not self._done and not self._canceled:
+            #     to = kwargs.get("timeout", None)
+            #     self._condition.wait(to)
+            #
+            #     if not self._done and not self._canceled:
+            #         self._condition.notify_all()
+                    # raise FutureTimeoutError("Wait timeout")
 
             self._condition.notify_all()
 
@@ -107,17 +111,29 @@ class Future(AbstractFuture):
 
     def cancel(self):
         with self._condition:
-            if not self._canceled or self._done:
-                self._future.cancel()
-                self._canceled = True
+            self._future.cancel()
+            # if not self._canceled or self._done:
+            #     self._future.cancel()
+            #     self._canceled = True
             self._condition.notify_all()
+
+    def is_done(self):
+        return self._done
 
     def done(self):
+        # self.exception()
         with self._condition:
-            if not (self._canceled or self._done):
-                self._condition.wait()
+            if self._future and not self._future.done():
+                try:
+                    self._future.result()
+                except Exception as e:
+                    self._exception = e
 
             self._condition.notify_all()
+
+    def exception(self):
+        if self._exception:
+            raise self._exception
 
 
 class SearchFuture(Future):